Fix keyword argument decorators
If a decorator is keyword-only, like `do |x: nil, y: nil|`,
then the following does nto work properly:

	factory.mydecorator(x: 1, y: 2)

Instead, both of them remain nil.
Ruby 3.0 changed how keyword arguments work
which broke some metaprogramming-heavy libraries
(this bug does not reproduce under 2.7).

My fix for this is to unpack a trailing hash
in the provided decorator arguments
and double-splat it, so this:

	o.instance_exec(*args, &block)

becomes something like this:

	o.instance_exec(*args[0..-2], **args[-1], &block)